python/fate_client/pyproject.toml
[tool.poetry]
name = "fate_client"
version = "1.10.0"
description = "Clients for FATE, including flow_client and pipeline"
authors = ["FederatedAI <contact@FedAI.org>"]
license = "Apache-2.0"
readme = "README.rst"

homepage = "https://fate.fedai.org/"
repository = "https://github.com/FederatedAI/FATE"
documentation = "https://fate.readthedocs.io/en/latest/?badge=latest"
keywords = ["FATE", "clients", "Federated Learning"]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Topic :: Software Development :: Testing",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

packages = [
    { include = "flow_client" },
    { include = "flow_sdk" },
    { include = "pipeline" }
]

[tool.poetry.dependencies]
python = "^3.8"
click = ">=7.1.2,<8.0.0"
loguru = ">=0.6.0"
poetry = ">=0.12"
pandas = ">=1.1.5"
requests = ">=2.24.0,<3.0.0"
requests_toolbelt= ">=0.9.1,<0.10.0"
"ruamel.yaml" = ">=0.16.10,<0.17.0"
setuptools = ">=50.0,<51.0"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
flow = "flow_client.flow:flow_cli"
pipeline = "pipeline.pipeline_cli:cli"

[build-system]
requires = ["poetry>=0.12", "setuptools>=50.0,<51.0"]
build-backend = "poetry.masonry.api"
